refactor prompt completion
This just splits the file completion.py into several files, one for each
Completer subclass.
