+++ alot/completion/__init__.py
@@ -1,0 +1,17 @@
+from .completer import Completer
+
+from .abooks import AbooksCompleter
+from .accounts import AccountCompleter
+from .argparse import ArgparseOptionCompleter
+from .command import CommandCompleter
+from .commandline import CommandLineCompleter
+from .commandname import CommandNameCompleter
+from .contacts import ContactsCompleter
+from .cryptokey import CryptoKeyCompleter
+from .multipleselection import MultipleSelectionCompleter
+from .namedquery import NamedQueryCompleter
+from .path import PathCompleter
+from .query import QueryCompleter
+from .stringlist import StringlistCompleter
+from .tag import TagCompleter
+from .tags import TagsCompleter
